chialisp_compiler/chialisp_compiler/requirements.txt
aiohttp==3.7.4
aiosqlite==0.17.0
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
bitstring==3.1.7
blspy==1.0.5
cffi==1.14.6
chardet==3.0.4
chia-blockchain==1.2.3
chia-dev-tools==1.0.3
chiabip158==1.0
chiapos==1.0.4
chiavdf==1.0.2
click==7.1.2
clvm==0.9.7
clvm-rs==0.1.8
clvm-tools==0.4.3
colorlog==5.0.1
concurrent-log-handler==0.9.19
cryptography==3.4.7
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
dnspython==2.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
jeepney==0.7.1
keyring==23.0.1
keyrings.cryptfile==1.3.4
multidict==5.1.0
packaging==21.0
pluggy==1.0.0
portalocker==2.3.2
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.15.1
pytimeparse==1.1.8
pytz==2021.1
PyYAML==5.4.1
SecretStorage==3.3.1
setproctitle==1.2.2
sortedcontainers==2.3.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
websockets==8.1
yarl==1.6.3
zipp==3.5.0
